Move constant files to standalone folder
