+++ src/const/degrees.ts
@@ -1,0 +1,14 @@
+export const degrees = [
+  "MD", // Doctor of Medicine
+  "PhD", // Doctor of Philosophy
+  "DO", // Doctor of Osteopathy
+  "MSW", // Master of Social Work
+  "PsyD", // Doctor of Psychology
+  "RN", // Registered Nurse
+  "LCSW", // Licensed Clinical Social Worker
+  "LMFT", // Licensed Marriage and Family Therapist
+  "NP", // Nurse Practitioner
+  "MSN", // Master of Science in Nursing
+  "MA", // Master of Arts (Psych)
+  "MEd", // Master of Education
+];
