+++ src/const/index.ts
@@ -1,0 +1,3 @@
+export * from "./cities";
+export * from "./degrees";
+export * from "./specialties";
