--- src/db/seed/const.ts
@@ -1,76 +1,0 @@
-export const degrees = [
-  "MD", // Doctor of Medicine
-  "PhD", // Doctor of Philosophy
-  "DO", // Doctor of Osteopathy
-  "MSW", // Master of Social Work
-  "PsyD", // Doctor of Psychology
-  "RN", // Registered Nurse
-  "LCSW", // Licensed Clinical Social Worker
-  "LMFT", // Licensed Marriage and Family Therapist
-  "NP", // Nurse Practitioner
-  "MSN", // Master of Science in Nursing
-  "MA", // Master of Arts (Psych)
-  "MEd", // Master of Education
-];
-
-export const specialties = [
-  "Bipolar",
-  "LGBTQ",
-  "Medication/Prescribing",
-  "Suicide History/Attempts",
-  "General Mental Health (anxiety, depression, stress, grief, life transitions)",
-  "Men's issues",
-  "Relationship Issues (family, friends, couple, etc)",
-  "Trauma & PTSD",
-  "Personality disorders",
-  "Personal growth",
-  "Substance use/abuse",
-  "Pediatrics",
-  "Women's issues (post-partum, infertility, family planning)",
-  "Chronic pain",
-  "Weight loss & nutrition",
-  "Eating disorders",
-  "Diabetic Diet and nutrition",
-  "Coaching (leadership, career, academic and wellness)",
-  "Life coaching",
-  "Obsessive-compulsive disorders",
-  "Neuropsychological evaluations & testing (ADHD testing)",
-  "Attention and Hyperactivity (ADHD)",
-  "Sleep issues",
-  "Schizophrenia and psychotic disorders",
-  "Learning disorders",
-  "Domestic abuse",
-];
-
-export const usCities = [
-  "New York",
-  "Los Angeles",
-  "Chicago",
-  "Houston",
-  "Phoenix",
-  "Philadelphia",
-  "San Antonio",
-  "San Diego",
-  "Dallas",
-  "San Jose",
-  "Austin",
-  "Jacksonville",
-  "Fort Worth",
-  "Columbus",
-  "Charlotte",
-  "San Francisco",
-  "Indianapolis",
-  "Seattle",
-  "Denver",
-  "Washington",
-  "Boston",
-  "El Paso",
-  "Nashville",
-  "Detroit",
-  "Oklahoma City",
-  "Portland",
-  "Las Vegas",
-  "Memphis",
-  "Louisville",
-  "Baltimore",
-];
